content_generator/pyproject.toml
[project]
name = "content-generator"
version = "1.0.0"
description = "Generate quiz questions using Tavily search and OpenRouter LLM"
requires-python = ">=3.11"
dependencies = [
    "tavily-python>=0.5.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
